Tic Tac Toe Game All set up to plays
Made all the function for user Authentication
Get all the game logic going with win or tie games
Got the updates of the game on every click of the user on the game Board
Created the buttons to access the fresh new game and get played information.
Did the README for other developer to understand the application better.

Comments are provided on each sections of code as header for what each function does.
Used console log on every event to test the code working.
Kept the game to the minimum simplest way i could to build the game.

Will work on improving my application as i keep learning new skills.